+++ lib/repositories/server_repository.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-enum IpState { init, notAuthorized, authorized }
+enum IpState { notAuthorized, authorized }
 
 class ServerRepository extends ChangeNotifier {}
